requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
black==21.6b0
click==8.0.1
Flask==2.0.1
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-dotenv==0.18.0
pytz==2021.1
regex==2021.7.6
six==1.16.0
toml==0.10.2
Werkzeug==2.0.1
